+++ app/hypatio/urls.py
@@ -19,6 +19,6 @@ urlpatterns = [
     re_path(r'^data-challenges/$', list_data_challenges, name='data-challenges'),
     re_path(r'^software-projects/$', list_software_projects, name='software-projects'),
     re_path(r'^healthcheck/?', include('health_check.urls')),
-    re_path(r'^groups/(?P<group_key>[^/]+)/?', GroupView.as_view(), name="group"),
-    re_path(r'^', index, name='index'),
+    re_path(r'^groups/(?P<group_key>[^/]+)/?$', GroupView.as_view(), name="group"),
+    re_path(r'^/?$', index, name='index'),
 ]